Move biip.gs1{ => _prefixes}.{GS1Prefix,GS1CompanyPrefix}
Breaking change
