+++ docs/api/gs1_prefixes.md
@@ -1,0 +1,3 @@
+# `biip.gs1_prefixes` - GS1 prefixes
+
+::: biip.gs1_prefixes
