+++ src/biip/gs1_prefixes/_company_prefix_trie.json.lzma
