+++ src/biip/sscc.py
@@ -45,7 +45,7 @@ from typing import Optional
 
 from biip import ParseError
 from biip.checksums import gs1_standard_check_digit
-from biip.gs1 import GS1CompanyPrefix, GS1Prefix
+from biip.gs1_prefixes import GS1CompanyPrefix, GS1Prefix
 
 
 @dataclass
