+++ tests/gs1/test_element_strings.py
@@ -6,12 +6,8 @@ import pytest
 
 from biip import ParseError
 from biip.gln import Gln
-from biip.gs1 import (
-    GS1ApplicationIdentifier,
-    GS1CompanyPrefix,
-    GS1ElementString,
-    GS1Prefix,
-)
+from biip.gs1 import GS1ApplicationIdentifier, GS1ElementString
+from biip.gs1_prefixes import GS1CompanyPrefix, GS1Prefix
 from biip.gtin import Gtin, GtinFormat
 from biip.sscc import Sscc
