+++ tests/test_gln.py
@@ -4,7 +4,7 @@ import pytest
 
 from biip import ParseError
 from biip.gln import Gln
-from biip.gs1 import GS1CompanyPrefix, GS1Prefix
+from biip.gs1_prefixes import GS1CompanyPrefix, GS1Prefix
 
 
 def test_parse() -> None:
